mixipy/requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-celery-results==2.0.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
kombu==5.0.2
Markdown==3.3.3
mysqlclient==2.0.3
prompt-toolkit==3.0.10
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
spotipy==2.16.1
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
